Time: 212 ms (23.85%), Space: 36.1 MB (23.33%) - LeetHub

+++ 2101-detonate-the-maximum-bombs/2101-detonate-the-maximum-bombs.cpp
@@ -33,7 +33,7 @@ public:
                 LL d2 = y2-y1;
                 LL d = d1*d1+d2*d2;
                 
-                if(r1*r1>=d){
+                if((LL)r1*r1>=d){
                     adj[i].push_back(j);
                 }
             }
